requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.14.0
transformers>=4.28.0
accelerate>=0.18.0
safetensors>=0.3.1
pillow>=9.5.0
flask>=2.3.2
flask-cors>=3.0.10
numpy>=1.24.3
tqdm>=4.65.0
requests>=2.30.0 